Electronic File Display and Entry

+++ SmartLibraryPortal/SmartLibraryPortal/uc_ElectronicFiles.cs
@@ -111,37 +111,48 @@ namespace SmartLibraryPortal
 
         private void button2_Click(object sender, EventArgs e)
         {
-
-            if (WebApiClient.EFileNameAvailability(txtFileName.Text))
+            if (ValidateChildren(ValidationConstraints.Enabled))
             {
-                bool isSuccess = false;
-                string EFileCopyDirectoryPath = ConfigurationSettings.AppSettings.Get("EFileCopyDirectoryPath").ToString() + "\\" + (cbCategory.SelectedItem as ComboBoxItem).Text;
-                string FileSQLPath = ConfigurationSettings.AppSettings.Get("EFileSQLPath").ToString() + "/" + (cbCategory.SelectedItem as ComboBoxItem).Text;
-
-                ElectronicFile eFile = new ElectronicFile()
+                if (WebApiClient.EFileNameAvailability(txtFileName.Text))
                 {
-                    FileName = txtFileName.Text,
-                    Path = FileSQLPath + (cbCategory.SelectedItem as ComboBoxItem).Text + "/" + Path.GetFileName(txtPath.Text),
+                    bool isSuccess = false;
+                    string EFileCopyDirectoryPath = ConfigurationSettings.AppSettings.Get("EFileCopyDirectoryPath").ToString() + "\\" + (cbCategory.SelectedItem as ComboBoxItem).Text;
+                    string FileSQLPath = ConfigurationSettings.AppSettings.Get("EFileSQLPath").ToString() + "/" + (cbCategory.SelectedItem as ComboBoxItem).Text;
 
-                    FileType = new ElectronicFileType()
+                    ElectronicFile eFile = new ElectronicFile()
                     {
-                        ElectronicFileTypeId = int.Parse((cbCategory.SelectedItem as ComboBoxItem).Value.ToString())
+                        FileName = txtFileName.Text,
+                        Path = FileSQLPath + "/" + Path.GetFileName(txtPath.Text),
+
+                        FileType = new ElectronicFileType()
+                        {
+                            ElectronicFileTypeId = int.Parse((cbCategory.SelectedItem as ComboBoxItem).Value.ToString())
+                        }
+                    };
+
+                    isSuccess = WebApiClient.AddElectronicFile(eFile);
+
+                    if (isSuccess)
+                    {
+                        MessageBox.Show((cbCategory.SelectedItem as ComboBoxItem).Text + " has been added");
+                        File.Copy(txtPath.Text, Path.Combine(EFileCopyDirectoryPath, Path.GetFileName(txtPath.Text)), true);
                     }
-                };
-
-                File.Copy(txtPath.Text, Path.Combine(EFileCopyDirectoryPath, Path.GetFileName(txtPath.Text)), true);
-
-
-                isSuccess = WebApiClient.AddElectronicFile(eFile);
-
-                if (isSuccess)
-                {
-                    MessageBox.Show("Eletronic file has been added");
                 }
-            }
-            else
-            {
-                MessageBox.Show((cbCategory.SelectedItem as ComboBoxItem).Text + " already exist by that name");
+                else
+                {
+                    MessageBox.Show((cbCategory.SelectedItem as ComboBoxItem).Text + " already exist by that name");
+                }
+
+                if (!this.DesignMode)
+                {
+                    list = GetPagedListAsync();
+
+                    btnPrevious.Enabled = list.HasPreviousPage;
+                    btnNext.Enabled = list.HasNextPage;
+                    dgElectronicFiles.DataSource = list.Select(o => new { Name = o.FileName, Type = o.FileType.Name, Format = o.Path.Substring(o.Path.LastIndexOf('.') + 1) }).ToList();
+                    lblPageCount.Text = string.Format("Page {0}/{1}", pageNumber, list.PageCount);
+                }
+
             }
         }
     
@@ -153,6 +164,36 @@ namespace SmartLibraryPortal
             if (opnfd.ShowDialog() == DialogResult.OK)
             {
                 txtPath.Text = opnfd.FileName;
+            }
+        }
+
+        private void txtFileName_Validating(object sender, CancelEventArgs e)
+        {
+            if (string.IsNullOrWhiteSpace(txtFileName.Text))
+            {
+                e.Cancel = true;
+                txtFileName.Focus();
+                errorEfile.SetError(txtFileName, "File Name should not be left blank!");
+            }
+            else
+            {
+                e.Cancel = false;
+                errorEfile.SetError(txtFileName, "");
+            }
+        }
+
+        private void txtPath_Validating(object sender, CancelEventArgs e)
+        {
+            if (string.IsNullOrWhiteSpace(txtPath.Text))
+            {
+                e.Cancel = true;
+                txtPath.Focus();
+                errorEfile.SetError(txtPath, "Please select a file");
+            }
+            else
+            {
+                e.Cancel = false;
+                errorEfile.SetError(txtPath, "");
             }
         }
     }
